+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- AS3 break loop in catch clause
+
 ## [13.0.0] - 2021-02-08
 ### Added
 - Graphviz graphs colorized